+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Fixed
 
 ### Changed
+- Removing unused CSS badge parts
 
 ### Breaking changes